# Speed up AntPathMatcher by doing prefix-match