@bclozel 2016-02-10
Thanks @soldierkam for this PR! I'd be happy to review more if you've got other ideas :+1: 

Indeed, I ran a few JMH benchmarks myself and found that this leads to a consistent performance boost for many cases. `AntPathMatcher` is a central piece of our infrastructure, so I'm not keen on backporting this change to the latest version of the 4.2.x line. This will be shortly available as a `4.3.0-BUILD-SNAPSHOT` version.

Merged as e77ff3c for [SPR-13913](https://jira.spring.io/browse/SPR-13913)